2docx/requirements.txt
google-cloud-storage
google-cloud-documentai
google-auth
python-docx